requirements.txt
Pillow
keras
tensorflow
sklearn
matplotlib
path.py
ipfml
cv2